java/src/main/java/com/yourtion/leetcode/daily/m11/d16/Solution.java
package com.yourtion.leetcode.daily.m11.d16;

import java.util.Arrays;

/**
 * 406. 根据身高重建队列
 *
 * @author Yourtion
 * @link https://leetcode-cn.com/problems/queue-reconstruction-by-height/
 */
public class Solution {
    public int[][] reconstructQueue(int[][] people) {
        Arrays.sort(people, (person1, person2) -> {
            if (person1[0] != person2[0]) {
                return person1[0] - person2[0];
            } else {
                return person2[1] - person1[1];
            }
        });
        int n = people.length;
        int[][] ans = new int[n][];
        for (int[] person : people) {
            int spaces = person[1] + 1;
            for (int i = 0; i < n; ++i) {
                if (ans[i] == null) {
                    --spaces;
                    if (spaces == 0) {
                        ans[i] = person;
                        break;
                    }
                }
            }
        }
        return ans;
    }
}
